+++ tencentcloud/essbasic/v20210526/models.go
@@ -1485,7 +1485,7 @@ type ChannelCreateEmbedWebUrlRequestParams struct {
 	// 用户自定义参数
 	// <ul>
 	// <li>目前仅支持EmbedType=CREATE_TEMPLATE时传入</li>
-	// <li>指定后，创建，编辑，删除模版时，回调都会携带该userData</li>
+	// <li>指定后，创建，编辑，删除模板时，回调都会携带该userData</li>
 	// <li>支持的格式：json字符串的BASE64编码字符串</li>
 	// <li>示例：<ul>
 	//                  <li>json字符串：{"ComeFrom":"xxx"}，BASE64编码：eyJDb21lRnJvbSI6Inh4eCJ9</li>
@@ -1543,7 +1543,7 @@ type ChannelCreateEmbedWebUrlRequest struct {
 	// 用户自定义参数
 	// <ul>
 	// <li>目前仅支持EmbedType=CREATE_TEMPLATE时传入</li>
-	// <li>指定后，创建，编辑，删除模版时，回调都会携带该userData</li>
+	// <li>指定后，创建，编辑，删除模板时，回调都会携带该userData</li>
 	// <li>支持的格式：json字符串的BASE64编码字符串</li>
 	// <li>示例：<ul>
 	//                  <li>json字符串：{"ComeFrom":"xxx"}，BASE64编码：eyJDb21lRnJvbSI6Inh4eCJ9</li>