+++ tencentcloud/teo/v20220901/models.go
@@ -3318,6 +3318,9 @@ type DescribeAccelerationDomainsRequestParams struct {
 	// <li>backup-origin： 按照备用源站地址进行过滤；</li>
 	// <li>domain-cname：按照 CNAME 进行过滤；</li>
 	// <li>share-cname：按照共享 CNAME 进行过滤；</li>
+	// <li>vodeo-sub-app-id：按照【 vodeo 子应用 ID】进行过滤；</li>
+	// <li>vodeo-distribution-range：按照【 vodeo 分发范围】进行过滤；</li>
+	// <li>vodeo-bucket-id：按照【vodeo 存储桶 ID】进行过滤；</li>
 	Filters []*AdvancedFilter `json:"Filters,omitnil" name:"Filters"`
 
 	// 可根据该字段对返回结果进行排序，取值有：
@@ -3355,6 +3358,9 @@ type DescribeAccelerationDomainsRequest struct {
 	// <li>backup-origin： 按照备用源站地址进行过滤；</li>
 	// <li>domain-cname：按照 CNAME 进行过滤；</li>
 	// <li>share-cname：按照共享 CNAME 进行过滤；</li>
+	// <li>vodeo-sub-app-id：按照【 vodeo 子应用 ID】进行过滤；</li>
+	// <li>vodeo-distribution-range：按照【 vodeo 分发范围】进行过滤；</li>
+	// <li>vodeo-bucket-id：按照【vodeo 存储桶 ID】进行过滤；</li>
 	Filters []*AdvancedFilter `json:"Filters,omitnil" name:"Filters"`
 
 	// 可根据该字段对返回结果进行排序，取值有：
@@ -8883,33 +8889,50 @@ type Origin struct {
 
 type OriginDetail struct {
 	// 源站类型，取值有：
-	// <li>IP_DOMAIN：IPV4、IPV6或域名类型源站；</li>
-	// <li>COS：COS源。</li>
-	// <li>ORIGIN_GROUP：源站组类型源站。</li>
-	// <li>AWS_S3：AWS S3对象存储源站。</li>
+	// <li>IP_DOMAIN：IPV4、IPV6 或域名类型源站；</li>
+	// <li>COS：腾讯云 COS 对象存储源站；</li>
+	// <li>AWS_S3：AWS S3 对象存储源站；</li>
+	// <li>ORIGIN_GROUP：源站组类型源站；</li>
+	//  <li>VODEO：云点播（混合云版）；</li>
+	// <li>SPACE：源站卸载，当前仅白名单开放；</li>
+	// <li>LB：负载均衡，当前仅白名单开放。</li>
 	OriginType *string `json:"OriginType,omitnil" name:"OriginType"`
 
-	// 源站地址，当OriginType参数指定为ORIGIN_GROUP时，该参数填写源站组ID，其他情况下填写源站地址。
+	// 源站地址，根据 OriginType 的取值分为以下情况：
+	// <li>当 OriginType = IP_DOMAIN 时，该参数为 IPv4、IPv6 地址或域名；</li>
+	// <li>当 OriginType = COS 时，该参数为 COS 桶的访问域名；</li>
+	// <li>当 OriginType = AWS_S3，该参数为 S3 桶的访问域名；</li>
+	// <li>当 OriginType = ORIGIN_GROUP 时，该参数为源站组 ID；</li>
+	// <li>当 OriginType = VODEO 时，如果 VodeoDistributionRange = ALL，则该参数为 "all-buckets-in-vodeo-application"；如果 VodeoDistributionRange = Bucket，则该参数为对应存储桶域名。</li>
 	Origin *string `json:"Origin,omitnil" name:"Origin"`
 
-	// 备用源站组ID，该参数在OriginType参数指定为ORIGIN_GROUP时生效，为空表示不使用备用源站。
+	// 备用源站组 ID，该参数仅在 OriginType = ORIGIN_GROUP 且配置了备源站组时会生效。
 	BackupOrigin *string `json:"BackupOrigin,omitnil" name:"BackupOrigin"`
 
-	// 主源源站组名称，当OriginType参数指定为ORIGIN_GROUP时该参数生效。
+	// 主源源站组名称，当 OriginType = ORIGIN_GROUP 时该参数会返回值。
 	OriginGroupName *string `json:"OriginGroupName,omitnil" name:"OriginGroupName"`
 
-	// 备用源站源站组名称，当OriginType参数指定为ORIGIN_GROUP，且用户指定了被用源站时该参数生效。
+	// 备用源站组名称，该参数仅当 OriginType = ORIGIN_GROUP 且配置了备用源站组时会生效。
 	BackOriginGroupName *string `json:"BackOriginGroupName,omitnil" name:"BackOriginGroupName"`
 
-	// 指定是否允许访问私有对象存储源站。当源站类型OriginType=COS或AWS_S3时有效 取值有：
+	// 指定是否允许访问私有对象存储源站，该参数仅当源站类型 OriginType = COS 或 AWS_S3 时会生效，取值有：
 	// <li>on：使用私有鉴权；</li>
 	// <li>off：不使用私有鉴权。</li>
 	// 不填写，默认值为off。
 	PrivateAccess *string `json:"PrivateAccess,omitnil" name:"PrivateAccess"`
 
-	// 私有鉴权使用参数，当源站类型PrivateAccess=on时有效。
+	// 私有鉴权使用参数，该参数仅当源站类型 PrivateAccess = on 时会生效。
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	PrivateParameters []*PrivateParameter `json:"PrivateParameters,omitnil" name:"PrivateParameters"`
+
+	// MO 子应用 ID
+	VodeoSubAppId *int64 `json:"VodeoSubAppId,omitnil" name:"VodeoSubAppId"`
+
+	// MO 分发范围，取值有： <li>All：全部</li> <li>Bucket：存储桶</li>	
+	VodeoDistributionRange *string `json:"VodeoDistributionRange,omitnil" name:"VodeoDistributionRange"`
+
+	// MO 存储桶 ID，分发范围(DistributionRange)为存储桶(Bucket)时必填
+	VodeoBucketId *string `json:"VodeoBucketId,omitnil" name:"VodeoBucketId"`
 }
 
 type OriginGroup struct {
@@ -8979,6 +9002,15 @@ type OriginInfo struct {
 
 	// 私有鉴权使用参数，当源站类型 PrivateAccess=on 时有效。
 	PrivateParameters []*PrivateParameter `json:"PrivateParameters,omitnil" name:"PrivateParameters"`
+
+	// MO 子应用 ID
+	VodeoSubAppId *int64 `json:"VodeoSubAppId,omitnil" name:"VodeoSubAppId"`
+
+	// MO 分发范围，取值有： <li>All：全部</li> <li>Bucket：存储桶</li>
+	VodeoDistributionRange *string `json:"VodeoDistributionRange,omitnil" name:"VodeoDistributionRange"`
+
+	// MO 存储桶 ID，分发范围(DistributionRange)为存储桶(Bucket)时必填	
+	VodeoBucketId *string `json:"VodeoBucketId,omitnil" name:"VodeoBucketId"`
 }
 
 type OriginProtectionInfo struct {
@@ -9402,6 +9434,10 @@ type Resource struct {
 	// 当前资源绑定的站点数量。
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	ZoneNumber *int64 `json:"ZoneNumber,omitnil" name:"ZoneNumber"`
+
+	// 资源标记类型，取值有：
+	// <li>vodeo：vodeo资源。</li>
+	Type *string `json:"Type,omitnil" name:"Type"`
 }
 
 type RewriteAction struct {