+++ CHANGELOG.md
@@ -1,3 +1,122 @@
+# Release v1.0.812
+
+## 商业智能分析 BI(bi) 版本：2022-01-05
+
+### 第 7 次发布
+
+发布时间：2023-12-06 01:10:35
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [CreateDatasourceCloud](https://cloud.tencent.com/document/api/590/98853)
+
+	* 新增入参：ClusterId
+
+* [CreateEmbedToken](https://cloud.tencent.com/document/api/590/73724)
+
+	* 新增入参：TicketNum
+
+* [ModifyDatasourceCloud](https://cloud.tencent.com/document/api/590/98850)
+
+	* 新增入参：ClusterId
+
+
+修改数据结构：
+
+* [EmbedTokenInfo](https://cloud.tencent.com/document/api/590/73726#EmbedTokenInfo)
+
+	* 新增成员：TicketNum
+
+
+
+
+## TDSQL-C MySQL 版(cynosdb) 版本：2019-01-07
+
+### 第 75 次发布
+
+发布时间：2023-12-06 01:34:06
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改数据结构：
+
+* [Package](https://cloud.tencent.com/document/api/1003/48097#Package)
+
+	* 新增成员：HistoryBindResourceInfos
+
+
+
+
+## 人脸核身(faceid) 版本：2018-03-01
+
+### 第 67 次发布
+
+发布时间：2023-12-06 01:49:35
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增接口：
+
+* [ImageRecognitionV2](https://cloud.tencent.com/document/api/1007/102203)
+
+
+
+## 游戏多媒体引擎(gme) 版本：2018-07-11
+
+### 第 36 次发布
+
+发布时间：2023-12-06 01:51:28
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+<font color="#dd0000">**预下线接口**：</font>
+
+* CreateAgeDetectTask
+* DescribeAgeDetectTask
+
+
+
+## 边缘安全加速平台(teo) 版本：2022-09-01
+
+### 第 67 次发布
+
+发布时间：2023-12-06 02:32:56
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改数据结构：
+
+* [OriginDetail](https://cloud.tencent.com/document/api/1552/80721#OriginDetail)
+
+	* 新增成员：VodeoSubAppId, VodeoDistributionRange, VodeoBucketId
+
+* [OriginInfo](https://cloud.tencent.com/document/api/1552/80721#OriginInfo)
+
+	* 新增成员：VodeoSubAppId, VodeoDistributionRange, VodeoBucketId
+
+* [Resource](https://cloud.tencent.com/document/api/1552/80721#Resource)
+
+	* 新增成员：Type
+
+
+
+
+## 边缘安全加速平台(teo) 版本：2022-01-06
+
+
+
 # Release v1.0.811
 
 ## Elasticsearch Service(es) 版本：2018-04-16